P10-S09: checked u64 vibefs block arithmetic and u64 sizes (F008)
map_block returns Result<Option<(usize, u32)>, Error>: it compares
extent ends in checked u64 and adds the physical block with checked_add
(Corrupt on overflow), and walks at most MAX_EXT extents.
split_replace_extent computes the right piece's log, phys and len in
checked u64 and converts with u32::try_from. truncate computes keep_blks
as u32::try_from(new.div_ceil(BLOCK)) and compares extent ends in checked
u64. Vol::read and Vol::write compute pos with checked_add and the file
block with u32::try_from (FileTooBig).

vibefs Node.size is u64 and node_from no longer saturates at 4 GiB, and
file_init's Walked.size is u64, so stat and SEEK_END are exact above
4 GiB. F008 leaves AGENTS.md rule 4's list, and VIBEFS.md §3 says the
block arithmetic is checked.

Tests: host node_size_above_4gib and block_math_near_u32_limit; in-guest
vibefs_seek_end_5gib.

Proves: vibefs_seek_end_5gib -- vibefs block arithmetic cannot overflow or truncate

+++ src/file_init.rs
@@ -37,7 +37,7 @@ struct Walked {
     vol: u8,
     ino: u32,
     kind: InodeKind,
-    size: u32,
+    size: u64,
     clu: u32,
     mode: u16,
     nlink: u32,
@@ -51,7 +51,7 @@ impl Walked {
             vol,
             ino: n.ino,
             kind: n.kind,
-            size: n.size,
+            size: u64::from(n.size),
             clu: n.clu,
             mode: if n.kind == InodeKind::Dir {
                 S_IFDIR_MODE
@@ -756,7 +756,7 @@ pub fn stat_path(path: &str) -> Result<Stat, FsError> {
         kind: node.kind,
         mode: node.mode,
         nlink: node.nlink,
-        size: node.size as u64,
+        size: node.size,
         atime: node.mtime,
         mtime: node.mtime,
         ctime: node.mtime,
@@ -832,7 +832,7 @@ pub fn vfs_attach(path: &str) -> Result<PathRef, FsError> {
             v.sb_of_path(pdir)?,
             node.ino,
             node.kind,
-            node.size as u64,
+            node.size,
             node.clu,
         )?;
         match v.fat_dcache(pdir, name, islot) {
